added: install and uninstall commands
