+++ src/modules/mod.rs
@@ -1,5 +1,7 @@
 pub mod cli;
 mod expand_archive;
+pub mod install_handler;
 pub mod ls_handler;
-pub mod use_handler;
 pub mod utils;
+pub mod use_handler;
+pub mod uninstall_handler;
